# Add SHA-512 hardware acceleration on ARMv8