--- a/doc/dev_ref/todo.rst
+++ b/doc/dev_ref/todo.rst
@@ -59,14 +59,14 @@ Utility Functions
 External Providers, Hardware Support
 ----------------------------------------
 
-* Add support ARMv8.4-A SHA-512, SHA-3, SM3 and RNG
+* Add support for ARMv8.4-A SHA-3, SM3 and RNG
 * Aarch64 inline asm for BigInt
 * /dev/crypto provider (ciphers, hashes)
 * Windows CryptoNG provider (ciphers, hashes)
 * Extend Apple CommonCrypto provider (HMAC, CMAC, RSA, ECDSA, ECDH)
 * Add support for iOS keychain access
 * POWER8 SHA-2 extensions (GH #1486 + #1487)
-* Add support VPSUM on big-endian PPC64 (GH #2252)
+* Add support for VPSUM on big-endian PPC64 (GH #2252)
 * Better TPM support: NVRAM, PCR measurements, sealing
 * Add support for TPM 2.0 hardware