… doc/hardware_acceleration.rst
… src/lib/hash/sha2_64/sha2_64_armv8/sha2_64_armv8.cpp