Adds support for ARMv8 hardware acceleration using instrinsics. It is heavily based on the existing SHA-256 implementation with changes to use the equivalent intrinsics for SHA-512.

Tested on a MacBook Pro with M2 Pro.

Baseline software:

```bash
$ ./botan speed --msec=3000 SHA-512
SHA-512 hash buffer size 1024 bytes: 489.480 MiB/sec (1468.440 MiB in 3000.000 ms)
```

ARMv8 intrinsics:

```bash
$ ./botan speed --msec=3000 SHA-512
SHA-512 hash buffer size 1024 bytes: 1307.809 MiB/sec (3923.428 MiB in 3000.000 ms)
```

Which means a factor 2,67 or 267% performance increase.

Even beats OpenSSL on the same machine (via Homebrew):

```bash
$ src/scripts/bench.py SHA-512
   INFO: Comparing Botan 3.3.0 with OpenSSL 3.2.0
   INFO: Running OpenSSL benchmark for SHA-512
   INFO: Running Botan benchmark for SHA-512
algo SHA-512 buf_size     16 botan    158735807 bps openssl     84757824 bps adv botan by 1.87
algo SHA-512 buf_size     64 botan    756406746 bps openssl    340762560 bps adv botan by 2.22
algo SHA-512 buf_size    256 botan    950167132 bps openssl    743652864 bps adv botan by 1.28
algo SHA-512 buf_size   1024 botan   1386804433 bps openssl   1238273024 bps adv botan by 1.12
algo SHA-512 buf_size   8192 botan   1565492078 bps openssl   1518950906 bps adv botan by 1.03
algo SHA-512 buf_size  16384 botan   1579776447 bps openssl   1556676608 bps adv botan by 1.01
```

Before, the baseline software was significantly slower:

```bash
$ src/scripts/bench.py SHA-512
   INFO: Comparing Botan 3.3.0 with OpenSSL 3.2.0
   INFO: Running OpenSSL benchmark for SHA-512
   INFO: Running Botan benchmark for SHA-512
algo SHA-512 buf_size     16 botan     69516433 bps openssl     82078000 bps adv openssl by 1.18
algo SHA-512 buf_size     64 botan    289047132 bps openssl    318918273 bps adv openssl by 1.10
algo SHA-512 buf_size    256 botan    389636626 bps openssl    711016210 bps adv openssl by 1.82
algo SHA-512 buf_size   1024 botan    504381686 bps openssl   1153344512 bps adv openssl by 2.29
algo SHA-512 buf_size   8192 botan    563204317 bps openssl   1458631111 bps adv openssl by 2.59
algo SHA-512 buf_size  16384 botan    580899115 bps openssl   1547288576 bps adv openssl by 2.66
```