@randombit 2023-12-31
Regarding

```

algo SHA-512 buf_size     16 botan    158735807 bps openssl     84757824 bps adv botan by 1.87
algo SHA-512 buf_size     64 botan    756406746 bps openssl    340762560 bps adv botan by 2.22
algo SHA-512 buf_size    256 botan    950167132 bps openssl    743652864 bps adv botan by 1.28
algo SHA-512 buf_size   1024 botan   1386804433 bps openssl   1238273024 bps adv botan by 1.12
algo SHA-512 buf_size   8192 botan   1565492078 bps openssl   1518950906 bps adv botan by 1.03
algo SHA-512 buf_size  16384 botan   1579776447 bps openssl   1556676608 bps adv botan by 1.01
```

you probably noticed that our advantage is confined to the smaller input sizes. I recently had to benchmark several (Rust) SHA-256 libraries using SHA-NI and noticed something similar; basically every implementation was fast for long messages, but for small messages there were huge swings in performance, and with OpenSSL always the worst. I believe this is due to differences in buffer management. This is a good sign that we're doing the right thing in this regard.