@randombit 2023-12-31
CI failures look relevent, but are confusing.

iOS:

```

  src/lib/hash/sha2_64/sha2_64_armv8/sha2_64_armv8.cpp:76:16: error: always_inline function 'vsha512h2q_u64' requires target feature 'sha3', but would be inlined into function 'compress_digest_armv8' that is compiled without support for 'sha3'
        STATE3 = vsha512h2q_u64(TSTATE1, STATE1, STATE0);
```

sha3?

Android:

```

  src/lib/hash/sha2_64/sha2_64_armv8/sha2_64_armv8.cpp:94:14: error: always_inline function 'vsha512su1q_u64' requires target feature 'sha3', but would be inlined into function 'compress_digest_armv8' that is compiled without support for 'sha3'
        MSG2 = vsha512su1q_u64(vsha512su0q_u64(MSG2, MSG3), MSG1, vextq_u64(MSG6, MSG7, 1));
```

sha3?

GCC

```
  In file included from src/lib/hash/sha2_64/sha2_64_armv8/sha2_64_armv8.cpp:10:
  /usr/lib/gcc-cross/aarch64-linux-gnu/11/include/arm_neon.h:32278:1: error: inlining failed in call to ‘always_inline’ ‘uint64x2_t vsha512h2q_u64(uint64x2_t, uint64x2_t, uint64x2_t)’: target specific option mismatch
  32278 | vsha512h2q_u64 (uint64x2_t __a, uint64x2_t __b, uint64x2_t __c)
        | ^~~~~~~~~~~~~~
  src/lib/hash/sha2_64/sha2_64_armv8/sha2_64_armv8.cpp:374:30: note: called from here
    374 |       STATE1 = vsha512h2q_u64(TSTATE1, STATE3, STATE2);
        |                ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~
```

Doesn't specify but I'm going to go ahead and guess it's sha3