@securitykernel 2024-01-01
> Oh right, instead of `armv8crypto` ISA flag you need `armv8sha512`. For whatever reason ARM bundles the SHA-512 and SHA-3 instructions in the same group.

That is it, thanks!